move event list from figure to axes
The mouse/keyboard events are bound to axes, not to the figure. This was the initial problem we wanted to solve. Reflect that by adding the events to the axes.

Before, if an axes was cleared, it was not obvious which of the figure-attached events/callbacks have to be cleared, too. Not clearing the callbacks results in stacking up of callbacks, calling them many (unneeded) times. Now, we easily can delete ax.UserData containing all the callbacks