proper view for new objects on a cleared axis
gfx.clearUiAxes3d-treated axes should properly scale for newly plotted objects. If the camera modes are set to manual BEFORE the objects are drawn, objects might not be visible at all while plotting. After the objects are drawn, the camera modes are set to manual on first interaction with the axes